Add support for PING frame
The protocol specifies a few things that could not be handled here due
to current Solicit APIs.

> If a PING frame is received with a stream identifier field value other
> than 0x0, the recipient MUST respond with a connection error (Section
> 5.4.1) of type PROTOCOL_ERROR.

and

> Receipt of a PING frame with a length field value other than 8 MUST be
> treated as a connection error (Section 5.4.1) of type
> FRAME_SIZE_ERROR.

When receiving a PING frame without the ACK flag set, the connection
replies with a PING frame having the ACK flag set. When an ACK is
received, nothing is done.

In addition to supporting the PING frame of the HTTP/2 protocol, a
method is exposed on the Async client to support sending PINGs as
desired by the client.