Async client responses are now async
The Async client response method previously returned a receiver which
the caller would have to block on. This resulted in wasted cycles if
some requests completed before the one being blocked on.

The design here allows associating user data with a request. The data is
provided back to the consumer in the request completion callback. This
design is a bit unfortunate since it requires the entire client being
generic to provide strong typing for the user data. Using `Any` wasn't
an option due to the desire for strong typing. The most attractive
alternative was proving a callback per request where such data could be
associated, but this would result in additional overhead.