Log reason async client service stops
All errors were previously being lost to the void. Errors are now
captured and logged using the `error!` macro. In the case of something
like the GOAWAY frame, the message is ideally printed as utf8 if it's a
valid utf8 string. The ConnectionError::debug_data now contains an enum
of either DebugData::Raw(Vec<u8>) or DebugData::Utf8(String) to support
this desired behavior.