ClientSession now stores connection settings
This makes it possible to use the provided ClientSession for clients
which are interested in pushing the limit of
SETTINGS_MAX_CONCURRENT_STREAMS.

This patch was designed to make use of the existing
`Session::new_settings` method. To achieve this, the ClientSession now
has an extra trait member of type `Settings`. The `Settings` trait
exists to support multiple implementations of settings managers. One is
provided out of the box, `SettingState`; this provided type simply
tracks settings in data members.

Only the SETTINGS_MAX_CONCURRENT_STREAMS setting is currently tracked.
If this is a reasonable way to move forward, we could add the rest of
the HTTP/2 connection settings to the types described above.