Return user data from async client on error
The async client has been updated with a mechanism to return user data
when an error is encountered. If the connection is closed with a
last_stream_id, user data from processed vs unprocessed streams is also
provided.

A new trait, ClientDelegate, makes this possible. This trait also
absorbed the done function of the async client.